+++ src/openfacefx/__init__.py
@@ -15,6 +15,8 @@ from .io_export import to_dict, write_json, write_csv
 from .export_unity import write_unity_anim, NAMING_PRESETS
 from .export_live2d import write_live2d_motion, lipsync_param_ids
 from .export_godot import write_godot_anim
+from .export_lip import (write_lip, lip_bytes, skyrim_mapping,
+                        SKYRIM_SLOT_ORDER)
 from .export_cues import (
     dominant_cues, write_rhubarb_tsv, write_rhubarb_xml, write_rhubarb_json,
     write_moho_dat, write_pgo,
@@ -45,6 +47,7 @@ __all__ = [
     "to_dict", "write_json", "write_csv",
     "write_unity_anim", "NAMING_PRESETS",
     "write_live2d_motion", "lipsync_param_ids", "write_godot_anim",
+    "write_lip", "lip_bytes", "skyrim_mapping", "SKYRIM_SLOT_ORDER",
     "dominant_cues", "write_rhubarb_tsv", "write_rhubarb_xml",
     "write_rhubarb_json", "write_moho_dat", "write_pgo",
     "Mapping", "Target",